Move HSJA into own directory
